tkf indel sim to separate file
